- fix reporting of the battery remaining time - fix showing the profiles combo (manual profile switching works, altho a bit flaky now, it's not correctly updated in the combo)
http://ktown.kde.org/~lukas/pics/kde46/battery-applet.png

svn path=/trunk/KDE/kdebase/workspace/; revision=1201016

+++ plasma/generic/applets/battery/battery.h
@@ -178,7 +178,7 @@ class Battery : public Plasma::PopupApplet
         int m_boxHoverAlpha;
         int m_numOfBattery;
         bool m_acAdapterPlugged;
-        int m_remainingMSecs;
+        qulonglong m_remainingMSecs;
 
         qreal m_labelAlpha;
         QPropertyAnimation *m_labelAnimation;
